+++ libraryManagement/Frmlogin.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace libraryManagement
+{
+    public partial class Frmlogin : Form
+    {
+        public Frmlogin()
+        {
+            InitializeComponent();
+        }
+
+        private void login_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void AdmPic_Click(object sender, EventArgs e)
+        {
+            
+           
+           
+
+
+
+        }
+
+        private void StdPic_Click(object sender, EventArgs e)
+        {
+            
+        }
+    }
+}
